Create part1.r
In general, problems consists of attributes of different data types. Such as binary, continuous, categorical, numerical, etc. In this code, I go through the process of showing the flexibility of the data types.

Binary (Yes/No) can be written as (0/1) and thus can be represented by numbers or characters.
Numbers (0,1...,1.456124) can be represented by numbers or characters. It is always possible to write "1.4324" as a string.
Strings can only be represented as characters.

Then, if we are interested in generating a matrix of one data type. The question remains: which data type? A very basic answer would be the strongest data type across your attributes and responses. For example, having Binaries and Numeric only, the chosen data type should be Numeric because we can write binaries as numeric (0/1). Another example if we have Binaries, Numeric and Strings, the chosen data type should be Strings (characters) because Numeric and Binaries can be written in characters.

However, it is still possible to create a mixed data type using data.frame in R. That is not very recommended because it would require taking cautions in all additional functions or scripts used. More chances of having logical errors that can not be caught easily.

This script was written by Abdallah Chehade in 2015 for Industrial Engineering Students.